trickster-agent/requirements.txt
# Core
httpx>=0.27,<1.0          # Async HTTP client for Moltbook API
pyyaml>=6.0               # Config parsing
python-dotenv>=1.0        # .env file loading
click>=8.1                # CLI framework

# AI
anthropic>=0.40           # Claude LLM for personality

# Storage
aiosqlite>=0.20           # Async SQLite for history

# Scheduling
apscheduler>=3.10,<4.0    # Cron-like scheduling

# Utilities
tenacity>=8.2             # Retry logic for API calls
